add eval

+++ train_contrastivepair.py
@@ -221,13 +221,14 @@ def validate(opts, model, val_loader, split, global_step):
 
     pat = re.compile(rf'{split}_results_(?P<step>\d+)_rank(?P<rank>\d+).csv')
     out_file = f'{opts.output_dir}/results/{split}_results_{global_step}.csv'
-    with open(out_file, 'wb') as g:
-        for f in glob.glob(f'{opts.output_dir}/results/{split}_results_{global_step}_rank*.csv'):
-            m = pat.match(os.path.basename(f))
-            rank = int(m.group('rank'))
-            if rank < opts.size:
-                shutil.copyfileobj(open(f, 'rb'), g)
-            os.remove(f)
+    if opts.rank == 0:
+        with open(out_file, 'wb') as g:
+            for f in glob.glob(f'{opts.output_dir}/results/{split}_results_{global_step}_rank*.csv'):
+                m = pat.match(os.path.basename(f))
+                rank = int(m.group('rank'))
+                if rank < opts.size:
+                    shutil.copyfileobj(open(f, 'rb'), g)
+                os.remove(f)
 
     sum(all_gather_list(opts.rank))
